handle banner

+++ VMS/Pages/Organization/Profile/EditOrganizationProfile.razor.cs
@@ -108,19 +108,25 @@ namespace VMS.Pages.Organization.Profile
 
         private async Task HandleSubmit()
         {
+            if (uploadFile is null)
+            {
+                await HandleInvalidSubmit();
+                return;
+            }
+
             isEditConfirm = !isEditConfirm;
             if (isEditSuccess)
             {
                 Logger.LogInformation("HandleValidSubmit called");
                 isErrorMessageShown = false;
                 isLoading = true;
-
+               
                 try
                 {
                     if (uploadFile is not null)
                     {
-                        string oldImageName = org.Avatar;
-                        org.Avatar = await UploadService.SaveImageAsync(uploadFile, OrgId);
+                        string oldImageName = org.Banner;
+                        org.Banner = await UploadService.SaveImageAsync(uploadFile, OrgId);
                         UploadService.RemoveImage(oldImageName);
                     }
                     await OrgService.UpdateOrgProfile(org, OrgId);
